refactor: mdx 관련 폴더 구조 개선
