+++ src/mdx/common/table-of-contents/table-of-contents.tsx
@@ -5,8 +5,8 @@ import { useEffect, useRef, useState } from 'react';
 
 import { Paragraph } from '@/components/ui/typography';
 import { cn } from '@/lib/utils';
-import { getHeaders, useActiveAnchor } from '../hook/use-toc';
 import type { MenuItem } from './toc';
+import { getHeaders, useActiveAnchor } from './use-toc';
 
 export function TableOfContents() {
   const [toc, setToc] = useState<MenuItem[]>([]);
