+++ src/mdx/common/table-of-contents/toc.d.ts
