+++ src/mdx/common/table-of-contents/use-toc.ts
@@ -1,7 +1,7 @@
 import { type RefObject, useEffect } from 'react';
 
 import { querySelectorAll } from '@/lib/dom';
-import type { HeadingLevel, LevelRange, MenuItem } from '../types/toc';
+import type { HeadingLevel, LevelRange, MenuItem } from './toc';
 
 // resolveHeaders 함수로부터 얻은 앵커 요소의 캐시된 목록
 const resolvedHeaders: { element: HTMLHeadingElement; link: string }[] = [];
