--- src/mdx/index.ts
@@ -1,3 +1,0 @@
-export * from './mdx';
-export { CustomMDX } from './mdx';
-export { TableOfContents } from './table-of-contents';
